# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * XenFi Wealth Management API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface RegisterRequest {
  name: string;
  email: string;
  /** @minLength 6 */
  password: string;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export interface User {
  id: number;
  name: string;
  email: string;
  isPremium: boolean;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export type AssetType = (typeof AssetType)[keyof typeof AssetType];

export const AssetType = {
  stocks: "stocks",
  crypto: "crypto",
  real_estate: "real_estate",
  cash: "cash",
} as const;

export interface Asset {
  id: number;
  userId: number;
  type: AssetType;
  name: string;
  value: number;
  purchaseValue: number;
  createdAt: string;
}

export interface CreateAssetRequest {
  type: AssetType;
  name: string;
  value: number;
  purchaseValue: number;
}

export interface UpdateAssetRequest {
  name?: string;
  value?: number;
  purchaseValue?: number;
}

export type ExpenseCategory =
  (typeof ExpenseCategory)[keyof typeof ExpenseCategory];

export const ExpenseCategory = {
  travel: "travel",
  dining: "dining",
  shopping: "shopping",
  bills: "bills",
  investment: "investment",
  other: "other",
} as const;

export interface Expense {
  id: number;
  userId: number;
  amount: number;
  category: ExpenseCategory;
  description: string;
  date: string;
  createdAt: string;
}

export interface CreateExpenseRequest {
  amount: number;
  category: ExpenseCategory;
  description: string;
  date: string;
}

export type LoanType = (typeof LoanType)[keyof typeof LoanType];

export const LoanType = {
  given: "given",
  received: "received",
} as const;

export type LoanStatus = (typeof LoanStatus)[keyof typeof LoanStatus];

export const LoanStatus = {
  pending: "pending",
  paid: "paid",
} as const;

export interface Loan {
  id: number;
  userId: number;
  name: string;
  type: LoanType;
  amount: number;
  dueDate: string;
  status: LoanStatus;
  createdAt: string;
}

export interface CreateLoanRequest {
  name: string;
  type: LoanType;
  amount: number;
  dueDate: string;
  status: LoanStatus;
}

export interface UpdateLoanRequest {
  name?: string;
  amount?: number;
  dueDate?: string;
  status?: LoanStatus;
}