# lib/api-zod/src/generated/types/asset.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * XenFi Wealth Management API
 * OpenAPI spec version: 0.1.0
 */
import type { AssetType } from "./assetType";

export interface Asset {
  id: number;
  userId: number;
  type: AssetType;
  name: string;
  value: number;
  purchaseValue: number;
  createdAt: Date;
}