# lib/api-zod/src/generated/types/assetType.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * XenFi Wealth Management API
 * OpenAPI spec version: 0.1.0
 */

export type AssetType = (typeof AssetType)[keyof typeof AssetType];

export const AssetType = {
  stocks: "stocks",
  crypto: "crypto",
  real_estate: "real_estate",
  cash: "cash",
} as const;