# lib/api-zod/src/generated/types/loanStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * XenFi Wealth Management API
 * OpenAPI spec version: 0.1.0
 */

export type LoanStatus = (typeof LoanStatus)[keyof typeof LoanStatus];

export const LoanStatus = {
  pending: "pending",
  paid: "paid",
} as const;