# lib/api-zod/src/generated/types/updateLoanRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * XenFi Wealth Management API
 * OpenAPI spec version: 0.1.0
 */
import type { LoanStatus } from "./loanStatus";

export interface UpdateLoanRequest {
  name?: string;
  amount?: number;
  dueDate?: Date;
  status?: LoanStatus;
}