# lib/db/src/schema/assets.ts
import { doublePrecision, integer, pgEnum, pgTable, serial, text, timestamp } from "drizzle-orm/pg-core";
import { createInsertSchema } from "drizzle-zod";
import { z } from "zod/v4";

export const assetTypeEnum = pgEnum("asset_type", [
  "stocks",
  "crypto",
  "real_estate",
  "cash",
]);

export const assetsTable = pgTable("assets", {
  id: serial("id").primaryKey(),
  userId: integer("user_id").notNull(),
  type: assetTypeEnum("type").notNull(),
  name: text("name").notNull(),
  value: doublePrecision("value").notNull(),
  purchaseValue: doublePrecision("purchase_value").notNull(),
  createdAt: timestamp("created_at").notNull().defaultNow(),
});

export const insertAssetSchema = createInsertSchema(assetsTable).omit({
  id: true,
  createdAt: true,
});

export type InsertAsset = z.infer<typeof insertAssetSchema>;
export type Asset = typeof assetsTable.$inferSelect;